refactor: set youtube embed as full url
